Hmin in racer_fml08_files was taken as integer, it should be real

+++ Testing/RacerNXS/racer_cfml08_files.f90
@@ -533,7 +533,7 @@ Module Racer_CFML_Files
 
            Snum%Scans(2)=nexus%scan_step
            Snum%wave=nexus%wave
-           Snum%hmin=nint(nexus%reflection)
+           Snum%hmin=nexus%reflection
            Snum%ub=nexus%ub
            Snum%icalc=1
            if(nexus%geometry == "NB")  Snum%icalc=2
